Add Util class in Frontend
